# Improve performance of businessCalendar's add()