--- a/business_calendar.go
+++ b/business_calendar.go
@@ -67,18 +67,30 @@ func (c businessCalendar) daysBetween(from, to date.Date) int {
 	switch start.Weekday() {
 	case time.Friday:
 		start = start.Add(2)
+
 	case time.Saturday:
 		start = start.Add(1)
-	case time.Monday, time.Tuesday, time.Wednesday, time.Thursday, time.Sunday:
+
+	case time.Monday,
+		time.Tuesday,
+		time.Wednesday,
+		time.Thursday,
+		time.Sunday:
 	}
 
 	// Shift end date to previous Friday if on Saturday or Sunday.
 	switch end.Weekday() {
 	case time.Saturday:
 		end = end.Add(-1)
+
 	case time.Sunday:
 		end = end.Add(-2)
-	case time.Monday, time.Tuesday, time.Wednesday, time.Thursday, time.Friday:
+
+	case time.Monday,
+		time.Tuesday,
+		time.Wednesday,
+		time.Thursday,
+		time.Friday:
 	}
 
 	// Compute raw day difference.