--- a/business_calendar_test.go
+++ b/business_calendar_test.go
@@ -338,34 +338,36 @@ func Test_businessCalendar_ConsistencyChecks(t *testing.T) {
 	})
 }
 
-func benchmarkShiftBusinessDays(b *testing.B, shift int, date date.Date) {
-	b.Helper()
+func BenchmarkShiftBusinessDays(b *testing.B) {
 	calendar := newBusinessCalendar()
-	for n := 0; n < b.N; n++ {
-		_ = calendar.add(date, shift)
+	startingDate := date.New(2017, time.November, 9)
+
+	for _, bc := range []struct {
+		name  string
+		shift int
+	}{
+		{
+			"add 17 days",
+			17,
+		},
+		{
+			"remove 17 days",
+			-17,
+		},
+		{
+			"add 7778 days",
+			17,
+		},
+		{
+			"remove 7778 days",
+			17,
+		},
+	} {
+		bc := bc
+		b.Run(bc.name, func(b *testing.B) {
+			for n := 0; n < b.N; n++ {
+				_ = calendar.add(startingDate, bc.shift)
+			}
+		})
 	}
-}
-
-func BenchmarkShiftBusinessDays_17(b *testing.B) {
-	date := date.New(2017, time.November, 9)
-	shift := 17
-	benchmarkShiftBusinessDays(b, shift, date)
-}
-
-func BenchmarkShiftminusBusinessDays_m17(b *testing.B) {
-	date := date.New(2017, time.November, 9)
-	shift := -17
-	benchmarkShiftBusinessDays(b, shift, date)
-}
-
-func BenchmarkShiftBusinessDays_7778(b *testing.B) {
-	date := date.New(2017, time.November, 9)
-	shift := 7778
-	benchmarkShiftBusinessDays(b, shift, date)
-}
-
-func BenchmarkShiftBusinessDays_m7778(b *testing.B) {
-	date := date.New(2017, time.November, 9)
-	shift := -7778
-	benchmarkShiftBusinessDays(b, shift, date)
 }