# Context

Similar to [edgelaboratories/tarantino#674](https://github.com/edgelaboratories/tarantino/pull/674)

This PR improves the performance of the function `add`.

# Content

Modification to `add` to avoid looping and calling itself. 

Benchmark added

# How we know it works

Tests pass. No new test added

# Benchmark results

### Intro

I considered four shift values `17`, `-17`, `7778` and `-7778`, referred to in the function names as `xxx_17`, `xxx_m17`, `xxx_7778`, `xxx_m7778`. The large values are irealistic and are used only for considerations of extremes.

### Results

version in master

```
BenchmarkShiftBusinessDays_17-8         	87126721	        11.93 ns/op	       0 B/op	       0 allocs/op
BenchmarkShiftminusBusinessDays_m17-8   	127907304	        9.362 ns/op	       0 B/op	       0 allocs/op
BenchmarkShiftBusinessDays_7778-8       	99965977	        11.94 ns/op	       0 B/op	       0 allocs/op
BenchmarkShiftBusinessDays_m7778-8      	100000000	        11.71 ns/op	       0 B/op	       0 allocs/op
```

version in this branch

```
BenchmarkShiftBusinessDays_17-8         	196057020	         6.129 ns/op	       0 B/op	       0 allocs/op
BenchmarkShiftminusBusinessDays_m17-8   	196296846	         6.138 ns/op	       0 B/op	       0 allocs/op
BenchmarkShiftBusinessDays_7778-8       	197015299	         6.086 ns/op	       0 B/op	       0 allocs/op
BenchmarkShiftBusinessDays_m7778-8      	195863832	         6.120 ns/op	       0 B/op	       0 allocs/op
```

### Discussion

The new version is slightly slower than tarantino's. I would assume this is because tarantino avoid using a date.Date. Also, the old version of calendar was not as bad as tarantino's old version. As a consequence, the improvement is not as extreme as observed in [edgelaboratories/tarantino#674](https://github.com/edgelaboratories/tarantino/pull/674).