@borisdavid 2022-06-28
Here are the existing benchmarks :

- `Benchmark_Calendar_Add`

On `master` :
![image](https://user-images.githubusercontent.com/79466627/176178845-9d532d0c-730b-4022-983c-bebb3131e516.png)

On this branch : 
![image](https://user-images.githubusercontent.com/79466627/176178629-d86fd25f-5028-41d6-b445-782d0714ee47.png)

- `Benchmark_Calendar_DaysBetween`

On `master` :
![image](https://user-images.githubusercontent.com/79466627/176178976-b72648f3-a3ea-49a1-a114-c59f6c508850.png)

On this branch : 
![image](https://user-images.githubusercontent.com/79466627/176179050-92942951-f1ba-4fc7-9123-208b9d197c3c.png)