@antonioiubatti93 2022-06-28
@borisdavid thx for [this](https://github.com/edgelaboratories/calendar/pull/24#issuecomment-1168661243) :wink: `Benchmark_calendar_add` sets are quite consequential (similar gain to the ones added in this PR), but the others don't show any significant impact. That's because the `DaysBetween` function still uses a for loop. I guess we could apply the same change there as well - in this case I'd find reasonable to do it here, as it's the same concept