@borisdavid 2022-06-28
I reworked the function `daysBetween` by trying out some other way to compute `businessDays` (hopeful not too creative :sweat_smile:). The `for`-loops contained in this function were quite marginals (not leading to a linear complexity) so the improvement will be minor compared to the function `add`. Here is the new benchmark for `daysBetween` : 
![image](https://user-images.githubusercontent.com/79466627/176206854-56792fac-cccc-445d-8d05-6f1fb62f35b1.png)